+++ lib/capybara_accessibility_audit/auditor.rb
@@ -8,7 +8,7 @@ module CapybaraAccessibilityAudit
 
     def audit!(method)
       if accessibility_audit_enabled && method.in?(accessibility_audit_after_methods) && javascript_enabled?
-        assert_no_accessibility_violations(**accessibility_audit_options)
+        audit_with_report_mode(**accessibility_audit_options)
       end
     end